+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New features:
 - Find controls now have history buffers, so up and down arrow will navigate previous searches.
 - Conversations now track the relationships between parent conversations and child conversations, so the mindspace conversations
   view now groups conversations under each other.
+- When opening an editor tab from a diff tab, the editor tab will be scrolled to the same position being shown in the diff.
 
 Bug fixes: